refactor/peer_list: split into multiple files
