+++ src/peer_list/mod.rs
@@ -6,23 +6,34 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
+mod membership_list;
+mod peer;
+mod peer_state;
+
+pub(crate) use self::membership_list::MembershipListChange;
+pub use self::peer_state::PeerState;
 #[cfg(all(test, feature = "mock"))]
 pub(crate) use self::snapshot::PeerListSnapshot;
+
+#[cfg(feature = "malice-detection")]
+use self::membership_list::MembershipListWithChanges;
+#[cfg(any(test, feature = "testing"))]
+use self::peer::Events;
+use self::peer::Peer;
 use error::Error;
 #[cfg(any(test, feature = "testing"))]
 use gossip::Graph;
 use gossip::{Event, EventIndex, IndexedEventRef};
 use hash::Hash;
-use id::{PublicId, SecretId};
+use id::SecretId;
 #[cfg(any(test, feature = "testing"))]
 use mock::{PeerId, Transaction};
 use network_event::NetworkEvent;
-use serialise;
 use std::collections::btree_map::{self, BTreeMap, Entry};
+#[cfg(any(test, feature = "testing", feature = "malice-detection"))]
 use std::collections::BTreeSet;
 use std::fmt::{self, Debug, Formatter};
 use std::iter;
-use std::ops::{BitOr, BitOrAssign};
 
 pub(crate) struct PeerList<S: SecretId> {
     our_id: S,
@@ -239,7 +250,7 @@ impl<S: SecretId> PeerList<S> {
     pub fn is_peer_membership_list_initialised(&self, peer_id: &S::PublicId) -> bool {
         self.peers
             .get(peer_id)
-            .map_or(false, |peer| !peer.membership_list.is_empty())
+            .map_or(false, |peer| !peer.membership_list().is_empty())
     }
 
     /// Same as `peer_membership_list_shapshot` except if there is a `Remove` at `event_index`,
@@ -458,285 +469,11 @@ impl PeerList<PeerId> {
     }
 }
 
-/// Peer state is a bitflag with these flags:
-///
-/// - `VOTE`: if enabled, the peer can vote, which means they are counted towards the supermajority.
-/// - `SEND`: if enabled, the peer can send gossips. For us it means we can send gossips to others.
-///           For others it means we can receive gossips from them.
-/// - `RECV`: if enabled, the peer can receive gossips. For us, it means we can receive gossips from
-///           others. For others it means we can send gossips to them.
-///
-/// If all three are enabled, the state is called `active`. If none is enabled, it's `inactive`.
-#[derive(Clone, Copy, PartialEq, Eq)]
-pub struct PeerState(u8);
-
-impl PeerState {
-    /// The peer is counted towards supermajority.
-    pub const VOTE: Self = PeerState(0b0000_0001);
-    /// The peer can send gossips.
-    pub const SEND: Self = PeerState(0b0000_0010);
-    /// The peer can receive gossips.
-    pub const RECV: Self = PeerState(0b0000_0100);
-
-    pub fn inactive() -> Self {
-        PeerState(0)
-    }
-
-    pub fn active() -> Self {
-        Self::VOTE | Self::SEND | Self::RECV
-    }
-
-    pub fn contains(self, other: Self) -> bool {
-        self.0 & other.0 == other.0
-    }
-
-    pub fn can_vote(self) -> bool {
-        self.contains(Self::VOTE)
-    }
-
-    pub fn can_send(self) -> bool {
-        self.contains(Self::SEND)
-    }
-
-    pub fn can_recv(self) -> bool {
-        self.contains(Self::RECV)
-    }
-}
-
-impl BitOr for PeerState {
-    type Output = Self;
-
-    fn bitor(self, rhs: Self) -> Self::Output {
-        PeerState(self.0 | rhs.0)
-    }
-}
-
-impl BitOrAssign for PeerState {
-    fn bitor_assign(&mut self, rhs: Self) {
-        self.0 |= rhs.0
-    }
-}
-
-impl Debug for PeerState {
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-        let mut separator = false;
-
-        write!(f, "PeerState(")?;
-
-        if self.contains(Self::VOTE) {
-            separator = true;
-            write!(f, "VOTE")?;
-        }
-
-        if self.contains(Self::SEND) {
-            if separator {
-                write!(f, "|")?;
-            }
-            separator = true;
-            write!(f, "SEND")?;
-        }
-
-        if self.contains(Self::RECV) {
-            if separator {
-                write!(f, "|")?;
-            }
-            write!(f, "RECV")?;
-        }
-
-        write!(f, ")")
-    }
-}
-
-#[derive(Debug)]
-pub struct Peer<P: PublicId> {
-    id_hash: Hash,
-    state: PeerState,
-    events: Events,
-    last_gossiped_event: Option<EventIndex>,
-    membership_list: BTreeSet<P>,
-    membership_list_changes: Vec<(usize, MembershipListChange<P>)>,
-}
-
-impl<P: PublicId> Peer<P> {
-    fn new(id: &P, state: PeerState) -> Self {
-        Self {
-            id_hash: Hash::from(serialise(id).as_slice()),
-            state,
-            events: Events::new(),
-            last_gossiped_event: None,
-            membership_list: BTreeSet::new(),
-            membership_list_changes: Vec::new(),
-        }
-    }
-
-    pub fn state(&self) -> PeerState {
-        self.state
-    }
-
-    pub fn events<'a>(&'a self) -> impl DoubleEndedIterator<Item = EventIndex> + 'a {
-        self.events.iter()
-    }
-
-    #[cfg(test)]
-    pub fn indexed_events<'a>(
-        &'a self,
-    ) -> impl DoubleEndedIterator<Item = (usize, EventIndex)> + 'a {
-        self.events.indexed()
-    }
-
-    pub fn events_by_index<'a>(&'a self, index: usize) -> impl Iterator<Item = EventIndex> + 'a {
-        self.events.by_index(index)
-    }
-
-    fn add_event(&mut self, index_by_creator: usize, event_index: EventIndex) {
-        self.events.add(index_by_creator, event_index);
-    }
-
-    #[cfg(test)]
-    fn remove_last_event(&mut self) -> Option<EventIndex> {
-        self.events.remove_last()
-    }
-
-    fn change_membership_list(&mut self, change: MembershipListChange<P>) {
-        if change.apply(&mut self.membership_list) {
-            self.record_membership_list_change(change);
-        }
-    }
-
-    fn record_membership_list_change(&mut self, change: MembershipListChange<P>) {
-        let index = self
-            .events
-            .indexed()
-            .rev()
-            .next()
-            .map(|(index, _)| index)
-            .unwrap_or(0);
-        self.membership_list_changes.push((index, change));
-    }
-
-    #[cfg(feature = "dump-graphs")]
-    pub(crate) fn membership_list(&self) -> &BTreeSet<P> {
-        &self.membership_list
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub(crate) enum MembershipListChange<P> {
-    Add(P),
-    Remove(P),
-}
-
-impl<P: Clone + Ord> MembershipListChange<P> {
-    fn apply(&self, peers: &mut BTreeSet<P>) -> bool {
-        match *self {
-            MembershipListChange::Add(ref peer_id) => peers.insert(peer_id.clone()),
-            MembershipListChange::Remove(ref peer_id) => peers.remove(peer_id),
-        }
-    }
-
-    #[cfg(feature = "malice-detection")]
-    fn unapply(&self, peers: &mut BTreeSet<P>) -> bool {
-        match *self {
-            MembershipListChange::Add(ref peer_id) => peers.remove(peer_id),
-            MembershipListChange::Remove(ref peer_id) => peers.insert(peer_id.clone()),
-        }
-    }
-
-    #[cfg(feature = "malice-detection")]
-    fn is_remove(&self) -> bool {
-        match *self {
-            MembershipListChange::Remove(_) => true,
-            MembershipListChange::Add(_) => false,
-        }
-    }
-}
-
-#[cfg(feature = "malice-detection")]
-type MembershipListWithChanges<'a, P> = (BTreeSet<P>, &'a [(usize, MembershipListChange<P>)]);
-
-#[derive(Debug)]
-struct Events(Vec<Slot>);
-
-impl Events {
-    fn new() -> Self {
-        Events(Vec::new())
-    }
-
-    fn add(&mut self, index_by_creator: usize, event_index: EventIndex) {
-        if let Some(slot) = self.0.get_mut(index_by_creator) {
-            slot.add(event_index);
-            return;
-        }
-
-        if index_by_creator != self.0.len() {
-            log_or_panic!("Peer events must be added sequentially");
-        }
-
-        self.0.push(Slot::new(event_index))
-    }
-
-    #[cfg(test)]
-    fn remove_last(&mut self) -> Option<EventIndex> {
-        if let Some(slot) = self.0.last_mut() {
-            if let Some(index) = slot.rest.pop() {
-                return Some(index);
-            }
-        } else {
-            return None;
-        }
-
-        self.0.pop().map(|slot| slot.first)
-    }
-
-    fn iter<'a>(&'a self) -> impl DoubleEndedIterator<Item = EventIndex> + 'a {
-        self.0.iter().flat_map(|slot| slot.iter())
-    }
-
-    fn indexed<'a>(&'a self) -> impl DoubleEndedIterator<Item = (usize, EventIndex)> + 'a {
-        self.0
-            .iter()
-            .enumerate()
-            .flat_map(|(index_by_creator, slot)| {
-                slot.iter()
-                    .map(move |event_index| (index_by_creator, event_index))
-            })
-    }
-
-    fn by_index<'a>(&'a self, index_by_creator: usize) -> impl Iterator<Item = EventIndex> + 'a {
-        self.0
-            .get(index_by_creator)
-            .into_iter()
-            .flat_map(|slot| slot.iter())
-    }
-}
-
-#[derive(Debug)]
-struct Slot {
-    first: EventIndex,
-    rest: Vec<EventIndex>,
-}
-
-impl Slot {
-    fn new(event_index: EventIndex) -> Self {
-        Self {
-            first: event_index,
-            rest: Vec::new(),
-        }
-    }
-
-    fn add(&mut self, event_index: EventIndex) {
-        self.rest.push(event_index)
-    }
-
-    fn iter<'a>(&'a self) -> impl DoubleEndedIterator<Item = EventIndex> + 'a {
-        iter::once(self.first).chain(self.rest.iter().cloned())
-    }
-}
-
 #[cfg(test)]
 pub(crate) mod snapshot {
     use super::*;
     use gossip::EventHash;
+    use id::PublicId;
 
     #[derive(Eq, PartialEq, Debug)]
     pub(crate) struct PeerListSnapshot<P: PublicId>(
